+++ framework/modules/facebook/controllers/facebookController.php
@@ -84,6 +84,7 @@ class facebookController extends expController {
                 'appId' => $params['config']['app_id'],
                 'secret' => $params['config']['app_secret'],
                 'cookie' => true,
+//                'fileUpload' => true     // this is important !
             ));
             $accesstoken = $facebook->getAccessToken();
             if (!empty($accesstoken)) {
@@ -107,6 +108,60 @@ class facebookController extends expController {
             } else {
                 // you're not logged in, the application will try to log in to get a access token
                 header("Location:{$facebook->getLoginUrl(array('scope' => 'photo_upload,user_status,publish_stream,user_photos,manage_pages'))}");
+    //            $dialog_url = "http://www.facebook.com/dialog/oauth?client_id=". $params['config']['app_id'] . "&redirect_uri=" . urlencode(URL_FULL) . "&scope=publish_stream,user_about_me,read_friendlists,offline_access,publish_actions,friends_photos,,user_photos,photo_upload,user_status,manage_pages". "&state=" . $_SESSION['fb_state'];
+    //            echo("<script> window.location.href='" . $dialog_url . "'</script>");
+                $status = gt('Permissions were not yet set on your Facebook page, please try again');
+                flash('error', $status);
+            }
+        }
+    }
+
+    public static function postEvent($params=array()) {
+        if (!empty($params)) {
+            // Include facebook class
+            require_once(BASE . "external/facebook-php-sdk-3.2.2/src/facebook.php");
+
+            // configuration
+//            $desc = 'Facebook constantly changes their SDK and methods for communicating with Facebook. The script in this post supports the latest Facebook authentication changes that will be implemented i October 2012.';
+//            $pic = 'http://blog.phpinfinite.com/wp-content/uploads/2012/11/post_to_facebook_from_php.jpg';
+//            $action_name = 'Go to PHP Infinite';
+//            $action_link = 'http://blog.phpinfinte.com';
+
+            $eventdate = new eventdate($params['id']);
+            $eventdate->event = new event($eventdate->event_id);
+
+            $facebook = new Facebook(array(
+                'appId' => $params['config']['app_id'],
+                'secret' => $params['config']['app_secret'],
+                'cookie' => true,
+//                'fileUpload' => true     // this is important !
+            ));
+            $accesstoken = $facebook->getAccessToken();
+            if (!empty($accesstoken)) {
+                try {
+                    $attachment = array(
+                        'access_token' => $accesstoken,
+                        "name"=>$eventdate->event->title,
+                        "description"=>expString::summarize($eventdate->event->body) . ' ' . expCore::makeLink(array('controller'=>$params['orig_controller'], 'action'=>'show','date_id'=>$eventdate->id)),
+                        "start_time"=>$eventdate->date + $eventdate->event->eventstart,
+                        "end_time"=>$eventdate->date + $eventdate->event->eventend,
+//                        "location"=>$location,
+//                        'description' => $desc,
+//                        'picture'=>$pic,
+//                        'actions' => json_encode(array('name' => $action_name,'link' => $action_link))
+                    );
+                    $result = $facebook->api("/".$params['config']['facebook_page']."/events", "post", $attachment);
+                    $facebookEventId = $result['id'];
+                    $status = gt('New Facebook Event posted');
+                    flash('message', $status);
+                } catch (FacebookApiException $e) {
+                    header("Location:{$facebook->getLoginUrl(array('scope' => 'photo_upload,user_status,publish_stream,user_photos,manage_pages,create_event'))}");
+                    error_log($e);
+                    flash('error', $e->getMessage());
+                }
+            } else {
+                // you're not logged in, the application will try to log in to get a access token
+                header("Location:{$facebook->getLoginUrl(array('scope' => 'photo_upload,user_status,publish_stream,user_photos,manage_pages,create_event'))}");
     //            $dialog_url = "http://www.facebook.com/dialog/oauth?client_id=". $params['config']['app_id'] . "&redirect_uri=" . urlencode(URL_FULL) . "&scope=publish_stream,user_about_me,read_friendlists,offline_access,publish_actions,friends_photos,,user_photos,photo_upload,user_status,manage_pages". "&state=" . $_SESSION['fb_state'];
     //            echo("<script> window.location.href='" . $dialog_url . "'</script>");
                 $status = gt('Permissions were not yet set on your Facebook page, please try again');