+++ framework/modules/twitter/controllers/twitterController.php
@@ -261,7 +261,7 @@ class twitterController extends expController {
     public static function postTweet($params=array()) {
         if (!empty($params)) {
             $post = new $params['model']($params['id']);
-            if (!empty($params['config']['consumer_key']) && !empty($post->body)) {
+            if (!empty($params['config']['consumer_key']) && !empty($post->title)) {
                 // create instance
                 if (expPermissions::check('manage', expCore::makeLocation($params['orig_controller'], $params['loc']))) {
                     $except_handler = 'twitter_exception_admin';
@@ -274,9 +274,36 @@ class twitterController extends expController {
                 $twitter->setOAuthToken($params['config']['oauth_token']);
                 $twitter->setOAuthTokenSecret($params['config']['oauth_token_secret']);
 
-                $twitter->statusesUpdate($params['config']['tweet_prefix'] . ' ' .expString::summarize($post->body) . ' ' . expCore::makeLink(array('controller'=>$params['orig_controller'], 'action'=>'show','title'=>$post->sef_url)));
+                $twitter->statusesUpdate($params['config']['tweet_prefix'] . ' ' . $post->title . ' ' . expCore::makeLink(array('controller'=>$params['orig_controller'], 'action'=>'show','title'=>$post->sef_url)));
                 restore_exception_handler();
                 flash('message', gt('New Tweet posted'));
+            }
+         }
+    }
+
+    /**
+     * Send an Event Tweet
+     */
+    public static function postEventTweet($params=array()) {
+        if (!empty($params)) {
+            $eventdate = new eventdate($params['id']);
+            $eventdate->event = new event($eventdate->event_id);
+            if (!empty($params['config']['consumer_key']) && !empty($eventdate->event->title)) {
+                // create instance
+                if (expPermissions::check('manage', expCore::makeLocation($params['orig_controller'], $params['loc']))) {
+                    $except_handler = 'twitter_exception_admin';
+                } else {
+                    $except_handler = 'twitter_exception';
+                }
+                set_exception_handler(array('twitterController', $except_handler));
+                $twitter = new expTwitter($params['config']['consumer_key'], $params['config']['consumer_secret']);
+                // set tokens
+                $twitter->setOAuthToken($params['config']['oauth_token']);
+                $twitter->setOAuthTokenSecret($params['config']['oauth_token_secret']);
+
+                $twitter->statusesUpdate($params['config']['tweet_prefix'] . ' ' . expDateTime::format_date($eventdate->date) . ' - ' . $eventdate->event->title . ' ' . expCore::makeLink(array('controller'=>$params['orig_controller'], 'action'=>'show','date_id'=>$eventdate->id)));
+                restore_exception_handler();
+                flash('message', gt('New Event Tweet posted'));
             }
          }
     }